# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/settings": {
      "get": {
        "operationId": "SettingsController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "SettingsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update settings",
        "tags": [
          "settings"
        ]
      }
    }
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "AdminKey": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SettingsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "Settings ID.",
            "default": 1
          },
          "notionalPercentage": {
            "type": "number",
            "example": 5,
            "description": "Notional percentage for settings.",
            "maximum": 100,
            "minimum": 1
          },
          "leverage": {
            "type": "number",
            "example": 10,
            "description": "Leverage for settings."
          },
          "takeProfitPercentage": {
            "type": "number",
            "example": 5,
            "description": "StopLoss percentage for settings.",
            "maximum": 100,
            "minimum": 1
          },
          "stopLossPercentage": {
            "type": "number",
            "example": 5,
            "description": "Notional percentage for settings.",
            "maximum": 100,
            "minimum": 1
          }
        },
        "required": [
          "id",
          "notionalPercentage",
          "leverage",
          "takeProfitPercentage",
          "stopLossPercentage"
        ]
      },
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {
          "notionalPercentage": {
            "type": "number",
            "example": 5,
            "description": "Notional percentage for settings.",
            "maximum": 100,
            "minimum": 1
          },
          "leverage": {
            "type": "number",
            "example": 10,
            "description": "Leverage for settings."
          },
          "takeProfitPercentage": {
            "type": "number",
            "example": 5,
            "description": "TakeProfit percentage for settings.",
            "maximum": 100,
            "minimum": 1
          },
          "stopLossPercentage": {
            "type": "number",
            "example": 5,
            "description": "Notional percentage for settings.",
            "maximum": 100,
            "minimum": 1
          }
        },
        "required": [
          "notionalPercentage",
          "leverage",
          "takeProfitPercentage",
          "stopLossPercentage"
        ]
      }
    }
  }
}